src layout
Breaks `daft.py` into submodules.
